requirements.txt
flask
gunicorn
icrawler
Pillow
